改本地资源为CDN资源，UI美化提升！

+++ admin.php
@@ -16,23 +16,17 @@
 		}	
 	}
 
-	// if(is_null($username)){
-	// 	exit;
-	// }
+	if(is_null($username)){
+		exit;
+	}
 	
 	// 删除POST
 	$get_id = trim($_POST['noteid']);
-
-	// if($username == NULL){
-	// 	$username = trim($_GET['username']);
-	// 	$_SESSION['username'] =  $username;
-	// }
 
 	if($get_id != NULL){
 		$sql = "DELETE FROM note WHERE noteid = $get_id";
 		mysqli_query($conn,$sql);
 	}
-	// mysqli_close($conn);
 
 	// 插入POST
 	$get_post = trim($_POST['post']);
@@ -73,21 +67,23 @@
 		<!--<div align="center"><a href="index.php">MessageBoard</a></div>-->
 		<title>MessageBoard Background</title>
 
-		<!-- Bootstrap core CSS -->
-		<link href="BT/docs/dist/css/bootstrap.min.css" rel="stylesheet">
+		<link href="index.css" rel="stylesheet">
 
-		<!-- IE10 viewport hack for Surface/desktop Windows 8 bug -->
-		<link href="BT/docs/assets/css/ie10-viewport-bug-workaround.css" rel="stylesheet">
+		<script src="https://cdn.bootcdn.net/ajax/libs/jquery/3.3.1/jquery.slim.min.js"></script>
 
-		<!-- Custom styles for this template -->
-		<link href="sticky-footer.css" rel="stylesheet">
-		<link href="jumbotron.css" rel="stylesheet">
+		<!-- 最新版本的 Bootstrap 核心 CSS 文件 -->
+		<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@3.3.7/dist/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
 
-		<!--<link rel="stylesheet" type="text/css" href="EU/themes/default/easyui.css">-->
-		<!-- <link rel="stylesheet" type="text/css" href="EU/themes/icon.css">
-		<link rel="stylesheet" type="text/css" href="EU/demo.css">
-		<script type="text/javascript" src="EU/jquery.min.js"></script>
-		<script type="text/javascript" src="EU/jquery.easyui.min.js"></script> -->
+		<!-- 可选的 Bootstrap 主题文件（一般不用引入） -->
+		<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@3.3.7/dist/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous">
+
+		<!-- 最新的 Bootstrap 核心 JavaScript 文件 -->
+		<script src="https://cdn.jsdelivr.net/npm/bootstrap@3.3.7/dist/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
+
+
+		<!--markdwwn需要-->
+		<script src="https://cdn.bootcss.com/jquery/1.11.3/jquery.min.js"></script>
+		<script src="js/editor.md-master/editormd.min.js"></script>
 
 		<!--markdwwn需要-->
 		<link rel="stylesheet" href="js/editor.md-master/css/editormd.css" />
@@ -96,19 +92,20 @@
 
 	<body>
 
-		<!--<nav class="navbar navbar-inverse navbar-fixed-top">-->
-		<nav>
+		<nav class="navbar navbar-inverse navbar-fixed-top">
 			<div class="container">
-					<a class="navbar-brand" style="font-size:32px;color:red;"><?php echo $username; ?>'s Admin Panel	</a>
+					<!-- <a href="index.php">HomePage</a> -->
+					<span></span>
+					<a class="navbar-brand" style="font-size:32px;">Hello,<?php echo $username; ?>!    <span>Wanna Say Something ?</a>
 					<div  align="right">
 						<form>
-							<button type="button" class="navbar-collapse" align="right" onClick="exit()">exit</button>
+							<button type="button" class="btn btn-danger btn-lg" align="right" onClick="exit()">exit</button>
 						</form>
 					</div>
 			</div>
 		</nav>
 		<hr/>
-		<!--<div align="center" style="font-size:32px;color:blue;">MessageBoardBackground</div>-->
+
 		<br/>
 			
 		<div align="center">
@@ -117,19 +114,22 @@
 					<tr>
 						<?php echo '<input type="hidden" name="username" value="' . $username . '"/>';	?>
 						<div id="test-editor">
-                    <textarea style="display:none;" name="mark">### 关于 Editor.md
-**Editor.md** 是一款开源的、可嵌入的 Markdown 在线编辑器（组件），基于 CodeMirror、jQuery 和 Marked 构建。</textarea>
-<!-- 第二个隐藏文本域，用来构造生成的HTML代码，方便表单POST提交，这里的name可以任意取，后台接受时以这个name键为准 -->
-
-                    <textarea class="editormd-html-textarea" name="post"></textarea>
-                </div>
+							<textarea style="display:none;" name="mark">### 关于 Editor.md**Editor.md
+** 是一款开源的、可嵌入的 Markdown 在线编辑器（组件），基于 CodeMirror、jQuery 和 Marked 构建。**
+> print("hello world !")
+</textarea>
+		<!-- 第二个隐藏文本域，用来构造生成的HTML代码，方便表单POST提交，这里的name可以任意取，后台接受时以这个name键为准 -->
+							<textarea class="editormd-html-textarea" name="post"></textarea>
+						</div>
 					</tr>
 					<tr>
-						<input type="submit" value="Submit" style="font-size:32px;">
+						<input type="submit" class="btn btn-success btn-lg" value="Submit" style="font-size:32px;">
 					</tr>
 				</table>
 			</form>
 		</div>
+
+		<hr/>
 
 		<div align="center" style="font-size:32px;color:blue;">Message Left By You !</div>
 		<br/>
@@ -140,27 +140,23 @@
 
 					$result = mysqli_query($conn, $sql);
 
-					// $row = mysqli_fetch_assoc($result);
-
+				
 					if ($result && mysqli_num_rows($result)) {					
 						//存在数据则循环将数据显示出来
-						echo '<table  align="center" border="1" width="95%" style="text-align:center;">';
-						echo '<tr><td>noteid</td><td>note</td><td>date</td><td>delete</td></tr>'; 
-						// while ($row = mysqli_fetch_assoc($result)) {
+						echo '<table class="table"  align="center" border="1" width="95%" >';
+						echo '<tr><td>noteid</td><td>note</td><td>date</td><td style="color:red;"><strong>delete?</strong></td></tr>'; 
 						while ($row = mysqli_fetch_assoc($result)) {
 							$id = $row['noteid'];
 							echo '<tr>';
-							echo '<td>' .$row['noteid'] .'</td>';
+							echo '<td style="text-align:center;">' .$row['noteid'] .'</td>';
 							echo '<td>' . $row['note'] . '</td>';
-							echo "<td>" . date("Y-m-d H:i:s",strtotime($row['date'])) . "</td>";
-							// echo '<td><a href="edit.php?id=' . $row['noteid'] . '">编辑</a></td>';
-							//	echo '<td><a href="delete.php?id='  . $row['noteid'] . '">删除</a></td>';
+							echo "<td style='text-align:center;'>" . date("Y-m-d H:i:s",strtotime($row['date'])) . "</std>";
 							echo '
-								<td>
+								<td style="text-align:center;">
 									<form action="#" method="POST">
 										<input type="hidden" name="noteid" value="' . $id . '"> 
 										<input type="hidden" name="username" value="' . $username . '"> 						
-                                        <button type="submit" class="btn btn-success" align="right">' . $id . '</button>					
+                                        <button type="submit" class="btn  btn-warning" align="right">' . $id . '</button>					
 									</form>
 								</td>';
 							echo '</tr>';
@@ -168,17 +164,17 @@
 						echo '</table>';
 
 					} else {
-						echo 'No Data';
+						echo '<div align="center" style="font-size:88px;color:red;">No Data ! <br>Wanna Say Something ?
+						</div>';
 					}
 				?>
 
 			</div>
 		</div>
 
-
-		<!--   <footer class="footer" style="font-size:18px">-->
-			<!--<p class="text-muted" align="center">Produced By<a style="text-decoration:none;" href="https://www.emperinter.info">emperinter</a>| <a style="text-decoration:none;" href="https://github.com/emperinter/MessageBoard">Github</a></p>-->
-		<!--   </footer>-->
+        <br/>
+        <br/>
+        <br/>
 
 		<footer class="footer mt-auto py-3">
 			<div class="container" align="center" style="text-decoration:none;font-size:22px;">
@@ -198,19 +194,7 @@
 		</script>
 
 
-		<!--markdwwn需要-->
-		<script src="https://cdn.bootcss.com/jquery/1.11.3/jquery.min.js"></script>
-		<script src="js/editor.md-master//editormd.min.js"></script>
 
-
-		<!-- Bootstrap core JavaScript
-		================================================== -->
-		<!-- Placed at the end of the document so the pages load faster -->
-		<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-		<script>window.jQuery || document.write('<script src="BT/docs/assets/js/vendor/jquery.min.js"><\/script>')</script>
-		<script src="BT/docs/dist/js/bootstrap.min.js"></script>
-		<!-- IE10 viewport hack for Surface/desktop Windows 8 bug -->
-		<script src="BT/docs/assets/js/ie10-viewport-bug-workaround.js"></script>		
 
 		<!-- 退出 -->
 		<script>
@@ -219,6 +203,7 @@
 				session_destroy();
 				// unset($_SESSION['username']);
 				// unset($_SESSION['password']);
+				mysqli_close($conn);
 				echo 'window.location.href = "index.php"';
 				?>
 			}
@@ -238,16 +223,5 @@
 </html>
 
 <?php
-
-	
-	// $get_id = trim($_GET['noteid']);
-	// $_SESSION['username'] =  trim($_GET['username']);
-
-	// if($get_id != NULL){
-	// 	$sql = "DELETE FROM note WHERE noteid = $get_id";
-
-	// 	mysqli_query($conn,$sql);
-	// }
 	// mysqli_close($conn);
-
 ?>
